+++ test/specs/openapi/stateful/test_engine.py
@@ -873,7 +873,7 @@ def test_multiple_incoming_link_without_override(app_factory):
     state_machine = schema.as_state_machine()
     assert (
         sum(len(operation.outgoing) for operation in state_machine._transitions.operations.values())
-        == schema.statistic.links.total
+        == schema.statistic.transitions.total
     )